bootcamp/requirements.txt
langchain
langgraph
ollama-python
n8n-client
pgvector
psycopg2-binary
dspy
